Permit zero pattern-cmd pairs before otherwise clause in 'match'
Previously, this would be a syntax error since no patterns to match
were specified:
	match foo {echo otherwise}
This now mirrors
	if {echo otherwise}
by behaving like
	match foo * {echo otherwise}
Since matching against * is guaranteed, it is rewritten to
	{echo otherwise}
rather than its equivalent if command, which would evaluate to the
same thing.

This commit also clarifies the general match command syntax in the man
page.