Remove abused getenv test
The abused getenv check was unreliable on some systems, and it was
added as a result of past mistakes with no need to be conditionally
defined after those mistakes were fixed.  It has now been removed, and
the replacement getenv() will always be used when compiling with
readline, and it will work as it's supposed to.