Implement setenv and putenv for readline
While unlikely to change in the near future, using an undocumented
function intended for Bash solely for updating the shell's definitions
of the LINES and COLUMNS environment variables is not a good idea.
LINES and COLUMNS environment variables are now set using a
replacement setenv or putenv function, and unsetenv is provided in
case it's needed.