Support for other readline implementations
To avoid a dependency on the Autoconf Archive, a more conservative
reimplementation of AX_LIB_READLINE is used.

One important difference is the fact that unless the library and the
readline.h header are both discovered, no HAVE_xxx macros are defined,
and history features will be disabled.

Another is the fact that instead of HAVE_LIBREADLINE, the test macro
is simply named HAVE_READLINE as LIBREADLINE is a misnomer when
linking with BSD libedit or Salz & Turner's editline.  The name also
does not indicate the correct headers are available.

Speaking of headers, the readline.h header is now required instead of
relying on manually declared extern variables and functions.
HAVE_READLINE_HISTORY will be declared if the optional history.h
header is discovered.  esconfig.h includes the correct headers and
contains redundant checks to ensure things work correctly.  It will
define READLINE and READLINE_HISTORY as necessary, though if READLINE
is undefined, so is READLINE_HISTORY.

Lastly, this change stashes auxiliary build files in the build-aux/
directory and macros in the m4/ directory to avoid cluttering the
source directory.