Replace mksignal shell script with mksignal.c
Rather than poking through header files, it is much easier to simply use
a C program to deal with signals since that method actually relies on
the preprocessor.

One of the goals of the mksignal script appears to have been replacing
the error messages for a signal like SIGIOT with equivalent messages
generated by another signal such as SIGABRT in the event that one signal
is defined in terms of another.  The new program orders the signals with
a preference for POSIX signal names, then other common/standard names,
followed lastly by any nonstandard/platform-specific names.  In the
event that two signals in the same group (i.e. POSIX/POSIX,
common/common, and platform/platform) are equivalent in value, the first
signal name when all signals in the group are sorted alphabetically is
the signal name chosen.  For example, on Linux systems where SIGSYS and
SIGUNUSED are equivalent, "SIGSYS" < "SIGUNUSED", so SIGSYS is chosen.

One last difference is that SIGPIPE and SIGINT will now print 'broken
pipe' and 'interrupted' respectively.  The old mksignal script would
assign empty strings as the messages for these signals, resulting in
nothing being printed, and this is the same behavior observed in Byron
Rakitzis's reimplementation of Plan 9 rc.  However, Plan 9 rc, Bash, and
many other shells notify users when a broken pipe occurs or when program
execution is interrupted, so we do the same.