Simplify signal handling configuration
Modern platforms typically have sigaction, and that is used if
available.

Other platforms may implement the a signal function with either System V
semantics or BSD semantics, and the old System V sighold and sigrelse
functions may also be implemented for compatibility.  We want BSD
semantics, but because both BSD signal and SysV sighold/sigrelse may
both be implemented, there is no way to guarantee that without a runtime
test using AC_RUN_IFELSE.

In short, checking whether both sighold and sigrelse exist is not enough
to determine whether we are using System V signal or BSD signal (and the
GNU C Library apparently implements both), so we always reset our custom
signal handler to be safe.