Replace setpgrp/setsid with setpgid
Previously, setsid would be used if available before falling back to
setpgid on Linux with GNU Libc.
If setsid was unavailable and the platform was not Linux with GNU Libc,
then setpgrp would have been used instead.

setpgrp is a legacy function usually only implemented for historical
compatibility in modern operating systems, so it has been dropped.

When a new session is created, the new process becomes the process group
leader of a foreground process, which is not what is wanted when using
background processes.
The last thing anybody wants is cat(1) or something else reading from
stdin to hang their shell unexpectedly instead of seeing SIGTTIN stop
the process, among other weird behaviors.

Instead, setpgid is the correct POSIX way to create a new process group
within an existing session while retaining the controlling terminal.
This way, terminal behavior of signals like SIGTTIN and SIGTTOU work
correctly when terminal I/O is attempted.