--enable-time to conditionally add $&time
Rather than rely on BUILTIN_TIME in esconfig.h, a new --enable-time
option was added.  This way, we can check whether to define BUILTIN_TIME
at configure time.

Autoconf complained about AC_FUNC_WAIT3, which is a run-time test, so it
was swapped for AC_CHECK_FUNC, which is a link-time test.  If a wait3
function is available, we can only assume it fills in the rusage
parameter.  If --disable-time (--enable-time=no) was used, the check is
still performed, but BUILTIN_TIME will not be defined.

When wait3 was unavailable, $&time used some trickery with times(3) to
get resource usage before execution, execute the code, then get resource
usage after execution and determine the difference.  Ignoring the fact
that times(3) should be replaced with getrusage, this did not always
work well, so now there is a dependency on wait3 for $&time.

Any rusage-related code is pointless if BUILTIN_TIME is not defined, so
instead of HAVE_WAIT3, the rusage-related code is defined only when
BUILTIN_TIME is nonzero.