Update code supporting $&limit primitive
Rather than a hard-coded linked list of suffixes, end pointers now exist
for the two types of suffixes that exist.  There are a couple of
additional branches to determine which end pointer to use based on the
given suffix pointer, but it increases the maintainability of the size
suffix list while also lowering memory usage.

More system-specific resources are available as well, and the Single
Unix Specification added RLIMIT_AS.  Some systems equivocate it with
RLIMIT_VMEM if both names are available, and the descriptions are indeed
nearly identical.  Solaris is one system that treats them differently.
OpenBSD does not have either one, and its lack of RLIMIT_AS is listed as
a bug in the manual for getrlimit/setrlimit.