Remove legacy checks
All modern POSIX systems have fcntl.h, unistd.h, and sys/time.h.  They
also have sigaction, sigsetjmp and friends, and sysconf.  A check has
been added to ensure that _SC_PAGESIZE is defined in the event that only
_SC_PAGE_SIZE is defined, but they should be equivalent on modern
implementations.

sys/ioctl.h is not defined by POSIX, and it was used for ioctl to set
the process group ID of the controlling terminal to the process group ID
of the process in $&newpgrp; this has since been replaced by tcsetpgrp.

setrlimit is classified as an X/Open Systems Interface, so it may not be
available, which is why we still check for it.

mmap was optional until POSIX.1-2008.  Instead of the runtime check
AC_FUNC_MMAP, which checks whether MAP_FIXED is supported, we simply
perform a compile check to determine whether MAP_PRIVATE is defined.  If
it compiles, then we don't care whether it actually works or not.  We
still fall back to /dev/zero in gc.c if MAP_ANONYMOUS is undefined.