Add .clang-format and .editorconfig files
Editors that support EditorConfig can take advantage of it now.

With the ClangFormat style file, we can use clang-format to consistently
format all source code.
There is unfortunately a problem with CatchException[If](...)
being indented incorrectly:

	ExceptionHandler
		... body ...
		CatchException(e)
		... catching code using e ...
	EndExceptionHandler

Instead, it should look like this:

	ExceptionHandler
		... body ...
	CatchException(e)
		... catching code using e ...
	EndExceptionHandler

If anybody gets this to work, I would be grateful.
The alternative is to overhaul the exception handling and GC.
That would be a lot of work, but it could be beneficial in the long run?