Make $&if catch exceptions in condition parameters
(wryun/es-shell#48)

Executing the following code with the -e flag would result in the shell
exiting:

    if {throw error foo bar} {echo ok} {echo fail}

To avoid this, one had to wrap every condition parameter inside a catch
block that effectively ignored the exception in favor of a false value.

With this change, $&if now catches exceptions in its condition
parameters on its own, so there is no longer a need for a catcher in
every condition where an exception might occur, and a condition that
raises an exception will not cause the shell to exit when -e is in use.