Replace getopt() with getopt_internal()
This replaces the native C library's getopt() functionality with a new
internal function getopt_internal() to combat the issue on some systems
with getopt() reordering the argument list.
For example, if the first line of the script is

	#!/path/to/es

and it is executed as

	./script -x

then what happens is

	/path/to/es ./script -x

is used.
es then uses getopt() to parse the command-line, which reorders the
command line into

	/path/to/es -x ./script

Switching to an internal implementation with POSIX semantics fixes this
issue.