Modify noreturn declarations
`noreturn` was previously a typedef for the nonstandard GNU extension
that considered functions returning `volatile void` as functions that
never returned to their caller.

C11 provided a standard equivalent in the form of the `_Noreturn`
type modifier, and C23 added a new attribute syntax, including a
`[[noreturn]]` attribute that may be placed before type qualifiers like
`extern`.
And GCC versions since a bit before 3.x actually provided their own
nonstandard attribute syntax, including `__attribute__((__noreturn__))`.

Unfortunately, C11 also mandated a `<stdnoreturn.h>` header that also
defines a `noreturn` macro expanding to `_Noreturn`, which would change
`$&noreturn` to `$&_Noreturn` if some system header included it.
To combat this issue, the `noreturn` macro is undefined just before
`PRIM(noreturn)`.

As a result of this whirlwind, the `noreturn` typedef has been changed
into the `es_noreturn` macro, expanding to one of the following:

* >=C23 - `[[noreturn]]`
* >=C11 - `_Noreturn`
* __GNUC__ >= 3 - `__attribute__((__noreturn__))`
* __GNUC__ < 3 - `volatile`
* else - `/*nothing*/`

The `void` return type has been added to the necessary prototypes and
function definitions as well since it was removed from the `es_noreturn`
macro for compatibility with C23.