Stores va_list pointers in Format
Both ISO C and POSIX requires every `va_start()` and `va_copy()`
invocation in a function to have a matching `va_end()` in the same
function.
To accomplish this, the `Format` type now stores a `va_list *`, so if a
function accepts parameter of type `Format *`, the `va_list *` can be
saved and later restored without worrying about undefined behavior.

This also avoids dangling pointers by not passing around `Format *`
everywhere and accidentally using a buffer allocated on the stack in
another function.