Avoid converting between void/function pointers
POSIX.1-2001 is the current target spec, which means casting function
pointers to/from `void *` is only possible on XSI systems.
This addresses the issue by storing primitives in an array instead of a
`Dict` and ensures that `setnoexport()` stores a non-NULL object
pointer in the `noexport` dict rather than a function pointer.

The array of primitives is generated by a new script `mkprims`.
`initprims()` starts by sorting all primitives by name, facilitating
efficient name-based lookup using a binary search algorithm.
`initprims()` finishes by ensuring that none of the primitives are
uninitialized, crashing the shell with a failed assertion otherwise.