Improve Sconv() code, add --enable-mbcs
Sconvp() would emit a single backslash escape and rely upon Sconv() for
concatenation and repeated attempts to decode the remaining bytes when
they were already decoded previously.
Rather than deal with that, it's easier to add a bit of concatenation
logic for backslash-escaped bytes and handle runs of backslash-escaped
bytes in the renamed Sconvp()->Sconvpsub() directly.

The autoconf code now includes a new --enable-mbcs option with a default
value of "check", which means you can now explicitly enable/disable
multi-byte support if desired, instead of configure automatically
enabling it if all requirements were satisfied.
Speaking of requirements, they weren't explicitly specified previously:

* wctype.h must exist and be compilable -- AC_CHECK_HEADER([wctype.h])
* mbrtowc() and mbstate_t must be declared properly -- AC_FUNC_MBRTOWC
* wchar_t must be able to store at least U+0000..U+10FFFF
    -- WCHAR_MAX >= 0x10ffffL

If any of those requirements are not met by an implementation,
support for multi-byte characters is unavailable.
That means if --enable-mbcs or --enable-mbcs=yes was explicitly
specified, configure will fail with an error message.