Add $&setlocale and fn-setlocale
`fn-setlocale` handles the shell semantics,
and `$&setlocale` invokes the C setlocale() function.

`$&setlocale` requires one or more locale arguments,
ignores empty strings, and returns the first locale that works,
or throws an exception if setlocale() never returned a non-null pointer.

Because of the settors for `LANG` and friends,
setlocale() does not need to be invoked in main().