Various locale handling changes
`$&setlocale` simply returns an empty list instead of throwing an
exception, making the `setlocale` builtin logic a bit simpler.
When `$&setlocale` returns an empty list, `setlocale` now throws an
exception instead.

`setlocale` also now temporarily unsets the settor for the given
variable and sets the variable to keep `setlocale VAR VALUE`
synchronized with the value of the dynamic variable `VAR`.
In the future, a `%setlocale` hook could prevent this.

Custom settors that catch the exception from the `setlocale` builtin are
now used to print the exception as an error message and return an empty
list, so the shell can start up properly.
In the future, it's possible those custom settors will be replaced with
`setlocale VAR` in main() before reading `.esrc` as they're really only
needed when reading the environment variables.

This also removes the custom setenv(), putenv(), and unsetenv()
implementations in favor of setting the LINES and COLUMNS variables
manually before each invocation of readline(), obtaining their new
(possibly unchanged) values using rl_get_screen_size().

The latter change necessitated the use of rl_initialize() in main() when
readline functionality will be used, which means rl_reset_screen_size()
will always be invoked, even before the first invocation of readline().
This fixed a bug where `LINES` and `COLUMNS` weren't always available
before the first invocation of `readline()`.

Some older versions/implementations of readline improperly handled
setlocale() returning a null pointer, so the value of `$LC_ALL` imported
from the environment is saved, `setlocale(LC_ALL, "C")` and
`rl_initialize()` are called, and `$LC_ALL` is restored.
If `$LC_ALL` is an empty list, this has the cascade effect of
effectively executing `$&setlocale LC_xxx $LC_xxx` for the `LC_xxx`
variables that are set and `$&setlocale LC_xxx $LANG 'C'` for the
`LC_xxx` variables that are unset.